test: Fix test compilation and expectations
- Add tempfile dev-dependency for test utilities
- Fix score type mismatch (f64 -> f32) in test helpers
- Fix Profile::default() to Profile::default_conservative()
- Update disk_analyzer test expectations (Code/SourceCode)
- Update grouping test expectations (pattern matching behavior)
- Fix verifier test to use CARGO_MANIFEST_DIR for reliable file path

All tests now pass: 31 passed; 0 failed

+++ crates/reclaim-core/src/cache.rs
@@ -457,7 +457,7 @@ mod tests {
     use super::*;
     use crate::candidate::{Action, TargetKind};
 
-    fn make_candidate(path: &str, score: f64) -> Candidate {
+    fn make_candidate(path: &str, score: f32) -> Candidate {
         Candidate {
             path: PathBuf::from(path),
             kind: TargetKind::Venv,
